bitboard merge

+++ game/src/main/java/org/toop/game/machinelearning/StateAction.java
@@ -1,9 +1,9 @@
 package org.toop.game.machinelearning;
 
 public class StateAction {
-    int[] state;
+    long[] state;
     int action;
-    public StateAction(int[] state, int action) {
+    public StateAction(long[] state, int action) {
         this.state = state;
         this.action = action;
     }
